feat(#66): 멤버 공부 화면 API 연동 - 기존 MemberId와 GroupId가 Int이던 것 Long으로 수정 - Paging3의  flow 데이터를 State로 넣는 건 좀 아닌 거 같아서 별도 변수로 작성

+++ app/src/main/java/com/ssafy/neegongnaegong/presentation/navigation/AppNavigation.kt
@@ -72,7 +72,7 @@ object AppNavigation {
             data object MakeVote : Studies
 
             @Serializable
-            data class Record(val groupId: Int, val memberId: Int) : Studies
+            data class Record(val groupId: Long, val memberId: Long) : Studies
         }
 
         @Serializable
